+++ server/app/src/controllers/consultations.controller.ts
@@ -12,7 +12,7 @@ import {
 import { ConsultationRepository } from '../repositories/consultation.repository';
 import { ConsultationDto } from '../dtos/consultations/consultation.dto';
 import { SaveConsultationDto } from '../dtos/consultations/save-consultation.dto';
-import { IdDto } from '../dtos/id.dto';
+import { IdDto } from '../dtos/id/id.dto';
 import { ConsultationMapper } from '../mappers/consultation.mapper';
 import { Consultation } from '../models/consultation.model';
 import { Connection } from 'typeorm';
