+++ server/app/src/controllers/groups.controllers.ts
@@ -11,7 +11,7 @@ import {
 import { GroupRepository } from '../repositories/group.repository';
 import { GroupDto } from '../dtos/groups/group.dto';
 import { SaveGroupDto } from '../dtos/groups/save-group.dto';
-import { IdDto } from '../dtos/id.dto';
+import { IdDto } from '../dtos/id/id.dto';
 import { Connection } from 'typeorm';
 import { GroupMapper } from '../mappers/group.mapper';
 import { Group } from '../models/group.model';
