+++ server/app/src/controllers/lecturers.controller.ts
@@ -11,7 +11,7 @@ import {
 import { LecturerRepository } from '../repositories/lecturer.repository';
 import { LecturerDto } from '../dtos/lecturers/lecturer.dto';
 import { SaveLecturerDto } from '../dtos/lecturers/save-lecturer.dto';
-import { IdDto } from '../dtos/id.dto';
+import { IdDto } from '../dtos/id/id.dto';
 import { Connection } from 'typeorm';
 import { LecturerMapper } from '../mappers/lecturer.mapper';
 import { Lecturer } from '../models/lecturer.model';
