+++ server/app/src/controllers/parents.controller.ts
@@ -11,7 +11,7 @@ import {
 import { ParentRepository } from '../repositories/parent.repository';
 import { ParentDto } from '../dtos/parents/parent.dto';
 import { SaveParentDto } from '../dtos/parents/save-parent.dto';
-import { IdDto } from '../dtos/id.dto';
+import { IdDto } from '../dtos/id/id.dto';
 import { Connection } from 'typeorm';
 import { ParentMapper } from '../mappers/parent.mapper';
 import { Parent } from '../models/parent.model';
