+++ server/app/src/controllers/students.controllers.ts
@@ -11,7 +11,7 @@ import {
 import { StudentRepository } from '../repositories/student.repository';
 import { StudentDto } from '../dtos/students/student.dto';
 import { SaveStudentDto } from '../dtos/students/save-student.dto';
-import { IdDto } from '../dtos/id.dto';
+import { IdDto } from '../dtos/id/id.dto';
 import { Connection } from 'typeorm';
 import { StudentMapper } from '../mappers/student.mapper';
 import { Student } from './../models/student.model';
